add comment preview (cannot control comment-text for preview)
